requirements.txt
pandas>=2.2
numpy>=1.26
PyYAML>=6.0
pytest>=8.0
matplotlib>=3.8
openpyxl>=3.1
exchange_calendars>=4.5
pandas_market_calendars>=4.4
